FTO-118 Error handling move to backend

+++ src/main/java/com/fotova/service/address/AddressService.java
@@ -2,6 +2,7 @@ package com.fotova.service.address;
 
 import com.fotova.dto.address.AddressDto;
 import com.fotova.entity.AddressEntity;
+import com.fotova.exception.NotFoundException;
 import com.fotova.repository.address.AddressRepositoryImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -24,16 +25,25 @@ public class AddressService {
 
     public AddressDto getAddressById(int id){
         AddressEntity addressEntity = addressRepository.findById(id);
+        if (addressEntity == null) {
+            throw new NotFoundException("Address with id " + id + " not found");
+        }
         return addressMapper.mpaToAddressDto(addressEntity);
     }
 
     public String deleteAddressById(int id){
+        if (addressRepository.findById(id) == null) {
+            throw new NotFoundException("Address with id " + id + " not found");
+        }
         addressRepository.updateClientAddressId(id);
         addressRepository.deleteById(id);
         return "Address has been deleted successfully for id : " + id;
     }
 
     public AddressDto updateAddress(AddressDto addressDto){
+        if (addressRepository.findById(addressDto.getId()) == null) {
+            throw new NotFoundException("Address with id " + addressDto.getId() + " not found");
+        }
         AddressEntity addressEntity = addressMapper.mapToAddressEntity(addressDto);
         AddressEntity addressEntityRes= addressRepository.update(addressEntity);
         return addressMapper.mpaToAddressDto(addressEntityRes);
